Split the New Simulation module into multiple smaller modules.
