+++ agents_playground/projects/template_file.py
@@ -1,0 +1,7 @@
+from typing import NamedTuple
+
+
+class TemplateFile(NamedTuple):
+  template_name: str
+  template_location: str
+  target_location: str
