+++ agents_playground/ui/new_sim_frame.py
@@ -1,19 +1,16 @@
 from __future__ import annotations
 
 from enum import StrEnum
-import importlib.metadata
-import os 
-from pathlib import Path 
-import re
-import shutil
-from string import Template
-import sys
 import traceback
-from typing import Any, NamedTuple
 
 import wx
 
 from agents_playground.loaders import now_as_string
+from agents_playground.projects.new_simulation_builder import NewSimulationBuilder, NewSimulationBuilderException
+from agents_playground.projects.simulation_template_options import SimulationTemplateOptions
+from agents_playground.ui.validators.cannot_be_empty_validator import CannotBeEmptyValidator
+from agents_playground.ui.validators.directory_must_exist_validator import DirectoryMustExistValidator
+from agents_playground.ui.validators.pattern_validator import PatternValidator
 
 ALLOWED_SIM_NAME_PATTERN    = r"^([a-z_])+$"
 NEW_SIM_GRID_BOARDER        = 5
@@ -35,230 +32,6 @@ NEW_SIM_NAME_FORMAT_ERROR        = "Only the lower case characters a-1 and the _
 NEW_SIM_TITLE_FORMAT_ERROR       = "The title cannot be empty."
 NEW_SIM_DESCRIPTION_FORMAT_ERROR = "The description cannot be empty."
 NEW_SIM_DIR_FORMAT_ERROR         = "The directory must exist."
-
-def keycode_to_value(keycode: int, shift_pressed: bool, caps_lock_on: bool) -> str:
-  """Given a keycode returns the related character.
-
-  Note: ASCII key codes use a single bit position between upper and lower case so 
-  x | 0x20 will force any key to be lower case.
-  
-  For example:
-    A is 65 or 1000001
-    32 -> 0x20 -> 100000
-    1000001 | 100000 -> 1100001 -> 97 -> 'a'
-  """
-  code = keycode if shift_pressed or caps_lock_on else keycode | 0x20
-  return chr(code)
-
-class InputErrorPopup(wx.PopupTransientWindow):
-  def __init__(self, parent: wx.Window, style: int, error_msg: str, font: wx.Font):
-    super().__init__(parent, style)
-    self._panel = wx.Panel(self)
-    self._panel.SetBackgroundColour('#FFB6C1')
-
-    error_label = wx.StaticText(self._panel, label='Input Error')
-    error_msg_component = wx.StaticText(self._panel, label=error_msg)
-    
-    error_label.SetFont(font.Bold())
-    error_label.SetForegroundColour('black')
-    error_msg_component.SetFont(font)
-    error_msg_component.SetForegroundColour('black')
-    
-    sizer = wx.BoxSizer(wx.VERTICAL)
-    sizer.Add(error_label, flag = wx.ALL, border = 5)
-    sizer.Add(error_msg_component, flag = wx.ALL, border = 5)
-    
-    self._panel.SetSizer(sizer)
-    sizer.Fit(self._panel)
-    sizer.Fit(self)
-    self.Layout()
-    
-class CannotBeEmptyValidator(wx.Validator):
-  def __init__(
-    self, 
-    error_msg: str, 
-    background_color: tuple[int,...], 
-    foreground_color: tuple[int,...]
-  ) -> None:
-    super().__init__()
-    self._error_msg = error_msg
-    self._original_background_color = background_color
-    self._original_text_color = foreground_color
-
-  def Clone(self):
-    return CannotBeEmptyValidator(
-      self._error_msg, 
-      self._original_background_color, 
-      self._original_text_color)
-  
-  def Validate(self, win):
-    """Called when the parent component does a self.Validate()"""
-    component = self.GetWindow()
-    value: str = component.GetValue()
-    valid = len(value.strip()) > 0
-    if valid:
-      component.SetBackgroundColour(self._original_background_color) 
-      component.SetForegroundColour(self._original_text_color) 
-      component.Refresh()
-    else:
-      error_popup = InputErrorPopup(
-        parent    = component, 
-        style     = wx.NO_BORDER, 
-        error_msg = self._error_msg,
-        font      = component.GetFont()
-      )
-
-      component_pos = component.ClientToScreen( (0,0) )
-      component_size =  component.GetSize()
-      error_popup.Position(component_pos, (0, component_size[1]))
-      error_popup.Popup()
-
-      component.SetForegroundColour('black')
-      component.SetBackgroundColour('pink')
-      component.SetFocus()
-      component.Refresh()
-    return valid
-  
-  def TransferToWindow(self):
-    return True
-
-  def TransferFromWindow(self):
-    return True 
-
-class DirectoryMustExistValidator(wx.Validator):
-  def __init__(
-    self, 
-    error_msg: str, 
-    background_color: tuple[int,...], 
-    foreground_color: tuple[int,...]
-  ) -> None:
-    super().__init__()
-    self._error_msg = error_msg
-    self._original_background_color = background_color
-    self._original_text_color = foreground_color
-
-  def Clone(self):
-    return DirectoryMustExistValidator(
-      self._error_msg, 
-      self._original_background_color, 
-      self._original_text_color)
-  
-  def Validate(self, win):
-    """Called when the parent component does a self.Validate()"""
-    component = self.GetWindow()
-    selected_path: str = component.GetPath()
-    print(selected_path)
-    valid = os.path.isdir(selected_path)
-    if valid:
-      component.SetBackgroundColour(self._original_background_color) 
-      component.SetForegroundColour(self._original_text_color) 
-      component.Refresh()
-    else:
-      error_popup = InputErrorPopup(
-        parent    = component, 
-        style     = wx.NO_BORDER, 
-        error_msg = self._error_msg,
-        font      = component.GetFont()
-      )
-
-      component_pos = component.ClientToScreen( (0,0) )
-      component_size =  component.GetSize()
-      error_popup.Position(component_pos, (0, component_size[1]))
-      error_popup.Popup()
-
-      component.SetForegroundColour('black')
-      component.SetBackgroundColour('pink')
-      component.SetFocus()
-      component.Refresh()
-    return valid
-  
-  def TransferToWindow(self):
-    return True
-
-  def TransferFromWindow(self):
-    return True 
-
-class PatternValidator(wx.Validator):
-  def __init__(
-    self, 
-    pattern: str, 
-    error_msg: str, 
-    background_color: tuple[int,...], 
-    foreground_color: tuple[int,...]
-  ) -> None:
-    super().__init__()
-    self._pattern = pattern
-    self._error_msg = error_msg
-    self._caps_lock_key_down: bool = False
-    self._original_background_color = background_color
-    self._original_text_color = foreground_color
-    self.Bind(wx.EVT_KEY_DOWN, self._handle_key_down)
-    self.Bind(wx.EVT_KEY_UP, self._handle_key_up)
-    self.Bind(wx.EVT_CHAR, self._handle_on_char)
-
-  def Clone(self):
-    return PatternValidator(
-      self._pattern, 
-      self._error_msg, 
-      self._original_background_color, 
-      self._original_text_color
-    )
-
-  def _handle_on_char(self, event: wx.KeyEvent) -> None:
-    """Only allow proper characters when typing."""
-    key_code: int = event.GetKeyCode()
-    key_value: str = keycode_to_value(key_code, event.ShiftDown(), self._caps_lock_key_down)
-    if key_code == wx.WXK_DELETE or re.match(self._pattern, key_value) is not None:
-      event.Skip()
-    return
-  
-  def _handle_key_down(self, event: wx.KeyEvent) -> None:
-    keycode = event.GetKeyCode()
-    if keycode == wx.WXK_CAPITAL:
-      self._caps_lock_key_down = True
-    event.Skip()
-    return 
-  
-  def _handle_key_up(self, event: wx.KeyEvent) -> None:
-    keycode = event.GetKeyCode()
-    if keycode == wx.WXK_CAPITAL:
-      self._caps_lock_key_down = False
-    event.Skip()
-    return 
-
-  def Validate(self, win):
-    """Called when the parent component does a self.Validate()"""
-    component = self.GetWindow()
-    value = component.GetValue()
-    match = re.match(self._pattern, value)
-    if match:
-      component.SetBackgroundColour(self._original_background_color) 
-      component.SetForegroundColour(self._original_text_color) 
-      component.Refresh()
-    else:
-      error_popup = InputErrorPopup(
-        parent    = component, 
-        style     = wx.NO_BORDER, 
-        error_msg = self._error_msg,
-        font      = component.GetFont()
-      )
-
-      component_pos = component.ClientToScreen( (0,0) )
-      component_size =  component.GetSize()
-      error_popup.Position(component_pos, (0, component_size[1]))
-      error_popup.Popup()
-
-      component.SetForegroundColour('black')
-      component.SetBackgroundColour('pink')
-      component.SetFocus()
-      component.Refresh()
-    return match is not None
-  
-  def TransferToWindow(self):
-    return True 
-
-  def TransferFromWindow(self):
-    return True 
 
 class SceneUOMOptions(StrEnum):
   METRIC = 'METRIC'
@@ -548,114 +321,3 @@ class NewSimFrame(wx.Frame):
           style   = wx.OK | wx.ICON_ERROR | wx.STAY_ON_TOP | wx.CENTRE
         )
         error_dialog.ShowModal()
-
-class SimulationTemplateOptions(NamedTuple):
-  simulation_name: str
-  simulation_title: str
-  simulation_description: str
-  parent_directory: str
-  creation_time: str
-  scene_uom_system: str
-  scene_distance_uom: str 
-  landscape_uom_system: str
-  tile_size_uom: str 
-
-  author: str
-  license: str
-  contact: str 
-  
-class TemplateFile(NamedTuple):
-  template_name: str
-  template_location: str
-  target_location: str
-
-# Note: All of the template data can use $var and ${var} for dynamic population.
-TEMPLATES = [
-  TemplateFile('__init__.py',         '$default_template_path', '$project_pkg'),
-  TemplateFile('cube_agent_def.json', '$default_template_path', '$project_pkg'),
-  TemplateFile('cube.obj',            '$default_template_path', '$project_pkg'),
-  TemplateFile('landscape.json',      '$default_template_path', '$project_pkg'),
-  TemplateFile('scene_test.py',       '$default_template_path', 'tests'),
-  TemplateFile('scene.json',          '$default_template_path', '$project_pkg'),
-  TemplateFile('scene.py',            '$default_template_path', '$project_pkg')
-]
-
-def populate_template(
-  path_to_template: Path, 
-  path_to_target: Path, 
-  template_inputs: dict[str, Any]
-) -> None:
-  scene_template: str = path_to_template.read_text()
-  scene_file = Template(scene_template).substitute(template_inputs)
-  path_to_target.write_text(scene_file)
-
-class NewSimulationBuilderException(Exception):
-  def __init__(self, *args: object) -> None:
-    super().__init__(*args)
-
-class NewSimulationBuilder:
-  def build(self, simulation_options: SimulationTemplateOptions) -> None:
-    """Assumes all validation has already happened and attempts to create a new simulation project."""
-    self._prevent_over_writing_existing_directories(simulation_options)
-    self._copy_template_directory(simulation_options)
-    self._rename_pkg_directory(simulation_options)
-    self._generate_project_files(simulation_options)
-    self._copy_wheel(simulation_options)
-
-  def _prevent_over_writing_existing_directories(self, simulation_options: SimulationTemplateOptions) -> None:
-    new_sim_dir = os.path.join(simulation_options.parent_directory, simulation_options.simulation_name)
-    if os.path.isdir(new_sim_dir):
-      error_msg = f'Cannot create the new simulation.\nThe directory {new_sim_dir} already exists.\nPlease specify a different project name.'
-      raise NewSimulationBuilderException(error_msg)
-    
-  def _copy_template_directory(self, simulation_options: SimulationTemplateOptions) -> None:
-    try:
-      new_sim_dir = os.path.join(simulation_options.parent_directory, simulation_options.simulation_name)
-      template_dir = os.path.join(Path.cwd(), 'agents_playground/templates/new_sim')
-      shutil.copytree(template_dir, new_sim_dir)
-    except Exception as e:
-      raise NewSimulationBuilderException('Failed to copy the template directory for the new simulation.') from e 
-
-  def _rename_pkg_directory(self, simulation_options: SimulationTemplateOptions) -> None:
-    try: 
-      package_dir = os.path.join(simulation_options.parent_directory, simulation_options.simulation_name, 'project_pkg')
-      project_pkg_dir = os.path.join(simulation_options.parent_directory, simulation_options.simulation_name, simulation_options.simulation_name)
-      Path(package_dir).rename(project_pkg_dir)
-    except Exception as e:
-      raise NewSimulationBuilderException('Failed to rename the package directory for the new simulation.') from e
-
-  def _generate_project_files(self, simulation_options: SimulationTemplateOptions) -> None:
-    try: 
-      template_inputs = simulation_options._asdict()
-      template_inputs['project_pkg'] = simulation_options.simulation_name
-      template_inputs['default_template_path'] = 'agents_playground/templates/base_sim_files'
-      new_project_dir = os.path.join(simulation_options.parent_directory, simulation_options.simulation_name)
-
-      for template in TEMPLATES:
-        hydrated_template_location: str = Template(template.template_location).substitute(template_inputs)
-        hydrated_template_name: str = Template(template.template_name).substitute(template_inputs)
-        hydrated_target: str = Template(template.target_location).substitute(template_inputs)
-        template_path: Path = Path(os.path.join(Path.cwd(), hydrated_template_location, hydrated_template_name))
-        target_path: Path   = Path(os.path.join(new_project_dir, hydrated_target, hydrated_template_name))
-        populate_template(template_path, target_path, template_inputs)
-    except Exception as e:
-      raise NewSimulationBuilderException('Failed to generate the project files for the new simulation.') from e
-
-  def _copy_wheel(self, simulation_options: SimulationTemplateOptions) -> None:
-    """Note
-    This is a temporary fix. The intent is to ultimately have the engine available 
-    on pypi.org as a Python package. The created project would then automatically 
-    install the package when the user runs the projects setup for doing unit testing.
-    """
-    try:
-      # Note: There is a thread around dealing with package versions for 
-      # poetry based packages.
-      # https://github.com/python-poetry/poetry/issues/273 
-      pkg_version = importlib.metadata.version('agents_playground')
-      engine_pkg = f'agents_playground-{pkg_version}-py3-none-any.whl'
-      wheel_path: Path = Path(os.path.join(Path.cwd(), 'dist', engine_pkg))
-      new_project_dir: Path = Path(os.path.join(simulation_options.parent_directory, simulation_options.simulation_name))
-      destination_path: Path = Path(os.path.join(new_project_dir, 'libs', engine_pkg))
-      shutil.copyfile(wheel_path, destination_path)
-    except Exception as e:
-      raise NewSimulationBuilderException("Failed to copy the engine's wheel while trying to set up the new simulation.") from e
